+++ src/main/java/com/CuteNekoDragon/Core/common/svogt/SVOGTRecipes.java
@@ -1,5 +1,6 @@
 package com.CuteNekoDragon.Core.common.svogt;
 
+import com.CuteNekoDragon.Core.SVOCore;
 import net.minecraft.data.recipes.FinishedRecipe;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Items;
@@ -11,7 +12,7 @@ import java.util.function.Consumer;
 public class SVOGTRecipes {
 
     public static void init(Consumer<FinishedRecipe> provider) {
-        SVOGTRecipeTypes.CHARKOAL_KILN_RECIPES.recipeBuilder("log_to_coal")
+        SVOGTRecipeTypes.CHARKOAL_KILN_RECIPES.recipeBuilder(SVOCore.id("log_to_coal"))
                 .inputItems(ItemTags.LOGS)
                 .outputItems(Items.COAL)
                 .duration(400)